Rename getScenarioLength() to readFile() and adjust the documentation.
This is in preparation for the complete implementation of readFile, which
will perform two scans of the file: the first to get bounds on the sizes of
the arrays needed (the current functionality of getScenarioLength), and the
second to actually read in all the data.

This also renames scan[Indep|Blocks]Line to scan[Indep|Blocks]Type.