Fix countStages() to accept also PERIODS lines with two values.
The SMPS standard allows the PERIODS IMPLICIT declaration, so there is
nothing wrong in having two values in a line other than the TIME line.
This fixes some tests for getStages, which were failing because they had a
line "PERIODS LP", which even if it doesn't conform to the standard, it
should not cause a failure.