Add the framework for the complete implementation of readFile().
This makes readFile do two scans of the stochastic file. Before starting
the quick scan, the position in the file is stored, to be retrieved when the
second scan needs to happen. This saves having to go through again the
lines with the names and the format type.

The proper reading happens in the new functions read[Indep|Block]Type, which
are mostly empty for now.