Have only one unit-test suite per class or filename.
This change the unit-test suites to have the name of the class in their
names. Each test suite can call tests for various functions. In the future
we'll want a way for each test to print its own name in case of failure, but
this is not possible right now.

This also moves the declarations of the unit-test suites from unit-tests.h
to unit-tests.cpp, so that when we add a new suite, only one file has to be
changed.