Add a pointer to an SmpsCore object to the SmpsTree class.
The SmpsTree class needs to refer to an initialized SmpsCore object, as it
has to access some of the information held by it. The SmpsTree::attachCore()
function sets the SmpsTree::core pointer: if the pointer is not set, then
readStocFile will not execute.

The code in test-SmpsTree.cpp calls attachCore() in order for the test to
work as expected.