Change the logic of the read[Smps|File|Stoc]File() functions.
The various reading routines can be called with an optional parameter.
Before, if they were called without an argument, they tried to read the
filename associated to their class, but the class filename was not updated.
Now, it they are called with an argument they update the class filename,
and that is the one that's read.

This change ensure that each class stores the filename of the file they are
actually reading.