Use strncmp() in the comparison for the ENDATA section.
When we check the ENDATA section, we are comparing the all buffer and not
only a token. So the buffer may contain additional whitespace after the
ENDATA token, but that should not raise a warning. The use of strncmp()
allows to compare only the first 6 letters of the buffer, and these are the
ones that should match ENDATA.