Reduce the coupling between the SmpsCore and SmpsTree classes.
The SmpsTree class can access the private members of SmpsCore because is a
friend. In some cases, it's not necessary to access directly the members of
SmpsCore, but their values can be retrieved through the public interface.

Therefore, in order to reduce the coupling, this moves convertPeriodNames()
to the protected interface, countRows() to the private interface, and
begPeriod[Row|Col] from SmpsTree to SmpsCore, where it makes more sense.