Fix bug in SmpsCore::readTimeFile().
The boolean value that records whether a match has been found between the row
names from the time file and those from the core file was not correctly reset
to false after a match was found. This was causing a segmentation fault if
the time file was declaring a row name that did not exist in the core file.