Add a link to the next node in order.
We need a fast way to traverse the tree so that we can loop though the nodes
in a specified order (breadth-first, depth-first, mixed). To do this we add
a new member to the node class to denote what the next node is going to be,
with the convention that NULL is used for the last node. This is equivalent
to creating a linked list of nodes.