Use the node representation of the stochastic event tree.
This changes the way we represent the event tree. Until now the event tree
was a collection of arrays for the parent, the number of children, and so on.
Now, instead, with the use of the Node class, the tree is represented by
a collection of nodes, and each node has links to the parent, its children,
and so on.

Since the stochastic file is read with rdstch(), the old array representation
is still generated, but it is then converted into nodes. In the future we
should avoid this intermediate step, since for a while we have double the
information (inside the nodes and inside the array), but this depends on
implementing a stochastic file reader that generates the information already
in the node representation.

While before we could traverse the tree by using a loop counter, now we have
to follow the links that each node has to the next one. So, the tree can
be traversed like a linked list.

The scenario changes are still in the old format, which is fine for now, as
it is not clear if it is convenient to store these changes inside the nodes,
as that could lead to data duplication.

The reordering routines of the SmpsOops class and the setting up of the
NodeInfo class are still not ready to be used.