Remove the mode parameter from the SmpsOops reordering routines.
The various reordering routines could work in two different modes which
refected into two different orderings of the rank corrector components.
However, only mode 1 was actually used.

This removes the ability of doing the other reordering, so that the function
signature is cleaner and the code slightly less complicated. Since the code
will have to be changed to work with the node representation, this commit is
just a way to allow us to see how get back to the previous state, if we ever
need to reimplement the different reordering.