Don't call SmpsCore::modifyCore() from ::readTimeFile().
If readTimeFile() calls modifyCore(), also the cplex interface will have
the slacks added to the formulation, which is not necessary. So we have to
call modifyCore() from SmpsOops::read(), and to do this, modifyCore() has
to be moved in the public interface.

This fixes also the printout on the number of slacks, so that it is printed
also when running in serial (the #endif WITH_MPI was misplaced).