Make SmpsTree::setNodeStarts() and  SmpsOops::orderNodes() work from any root.
For warmstart, we may have different trees associated to the same problem,
and these need have the correct row and column indices for the deterministic
equivalent, and be reordered correctly.

Therefore, we add an argument to setNodeStart() and orderNodes() to specify
which is the root node of the tree they have to work on.