Print also the next node from Node::print().
This allows to print the next node in the linked list, with a value of -1
if the node is the one in the list (so that it doesn't have a next node).

This also prints out the scenario number starting from 1 rather than 0, for
consistency with what was done in the old code.