Work out whether the problem has upper bounds and deal with them.
To know whether a problem has some upper bounds set, we need to add a member
variable to SmpsCore and the corresponding getter.

If the problem is bounded, then explicitly allocate the vectors vs and vw.
While these would have been allocated anyway by Oops, this is an essential
step to make the smps interface deal with upper bounds during warmstart.