Make SmpsOops::generateSmps() be able to deal with any given root node.
It is essential for warmstart that we generate the deterministic equivalent
according to a given event tree. Therefore, generateSmps() has to accept a
root node as an argument, and use it in the generation process.

We store this root node in the SmpsReturn class, so that it can be accessed
by the other routines called by generateSmps().