Move countNonzeros() and nzPeriod to the Smps class.
This change is motivated by the fact that we will need to be able to count
the number of nonzeros depending on a specific event tree. While this is
possible right now in the SmpsTree class by indicating a root node, in the
future we will want to do it by indicating an SmpsTree object from the Smps
class.

This also change SmpsCore::countNzPeriodBlocks() to return the nzPeriod array
that is now a member of the Smps class.