Split the SmpsStoc class from SmpsTree.
In order to make the SmpsTree class contain only the information related to
the tree, we split off the SmpsStoc class to contains the information on the
scenario changes.

This changes a bit the overall design. The Smps class derives from SmpsStoc
and contains an SmpsTree corresponding to the event tree read from the
stochastic file. The Smps class provides some methods to access and update
information stored in the SmpsTree. In particular, it now provides
setNodeStarts().

Some of the methods that were previously working from a root node, now will
work from an SmpsTree. The reduced event tree of the SmpsOops class is now
an SmpsTree rather than a node.