Don't crash if Node::addChild() is called on a NULL node.
Rather than trying to reference a NULL node, we return immediately and so we
avoid a crash.

This turns out to be useful when we try to add nodes to a parent node within
a loop, where we may end up to the root node and try to add a child to the
root's parent node (which is NULL by definition).