Add a switch for aggregation and rework the meaning of useWarmstart().
This adds a command line option to activate stage aggregation. Since this is
an alternative warmstart strategy, we need a way to differentiate warmstart
by scenario reduction and by stage aggregation.

We now store the value of the "-w" option inside _useReduction, and change
the meaning of _useWarmstart to be 1 if any warmstart strategy has been
requested, and 0 otherwise.