Regenerate the column names in SmpsCore::modifyCore().
In modifyCore() we add the slack variables, and therefore we need to rebuild
the vector of column names to take them into account.

We need to remove the deletion of the '\0' character, because this leaves
array of names unterminated, which creates problems inside convertNames().
The removal of this line doesn't seem to affect anything, and I'm not sure
why it was there in the first place.