Build the row and column names only if needed.
The row and column names for the deterministic equivalent are needed only
if the mps file has to be printed out. Hence, they should be built only in
this case.

This adds a member to the Smps class to know whether the mps names should be
built, and a method to switch this on. This method is called when the option
to print the mps file is read.