Move the call to SmpsCore::modifyCore() into Smps::read().
In order to get the correct indices of the scenario changes in the matrix,
the slack variables need to be added before the stochastic file is read.
This adds a default parameter to Smps::read() through which we can switch
on the call to modifyCore() if it's needed to add slacks (for Oops).