Fix errors with string delimiters in SmpsCore::modifyCore().
The first change adds back a line that was removed a few days ago. We really
have to delete the '\0' that is added by sprintf() when we write the name
of the slack, because otherwise we would only print this array up to the
first slack's name. The second change adds a '\0' to terminate the character
array, so that if we print it we do not overflow the string.