Ensure that we add as many slack variables as necessary.
If the time file declares the beginning of the first period in the wrong
position and thus leaving out a row, then wronlgy we wouldn't add a slack
for that row.

Now we assert that we add as many slacks as we can compute by looking at
the core file.