Apply the scenario corrections to the above-diagonal elements.
This allows to apply the scenario corrections to above-diagonal elements
instead of crashing or exiting.

It still needs some testing, as it seems that in xi-5 we do not generate the
correct matrix: changes in an above-diagonal element in one scenario are not
applied for the scenarios that descend from that one.