Fix copyLinkingBlocks() for problems that don't have staircase structure.
The previous code was assuming to have staircase structure, that is a link
between at most two consecutive stages. This change allows to deal with
problems that display links between not consecutive stages.

The elements in non consecutive stages have to be copied as many times as
there are nodes in their row period. This is accomplished by recursion on
the children of the current node.