Report an error if the scanning of the stochastic file fails.
The scanning of the stochastic file fails if it cannot find the NAME block.
In this case, stocType is still set to 0, so none of the scanning routines
is invoked. If this happens, we have to return early because the number of
nodes and scenarios has not been computed, and we would fail in rdstch().

The fix is to set the return value to -1 rather than to stocType, so that
we trigger the early return. In this case, we report an error message.