Allocate enough space for the name of the matrix blocks.
The character array for the names needs to be correctly allocated before
passing it to sprintf(). The size 15 has been chosen so that it is enough
for names such as Border[999999], which should be plenty.

I don't know why valgrind did not spot this when compiling with gcc, but
it was wrong at it didn't work with the intel compiler.