Fix generateSmps() when aggregation is required.
The rewriting of the way the blocks are filled in generateSmps() broke the
warmstart by aggregation. Since in the aggregation case a node may span
more than one period, to determine the start of the next period we have to
take into account node->nLevels() instead of assuming that it's 1.