Change the mapping of the nodes in SmpsOops::reduceScenarios().
Up to now we would traverse the tree in order and choose the nodes to be
put in the reduced tree sequentially. While this is fine for multistage
problems, it doesn't really work for two stage problems.

The issue is that for each subtree we were mapping all nodes to only one
reduced tree scenario. For two stage problems, there is only one subtree
(that corresponds to the whole tree itself), and so we were not able to
initialise the complete tree solution by using more than one reduced tree
scenario.

The problem was that we were creating a mapping between complete tree and
reduced tree only for the nodes that were put in the reduced tree, while
for the other nodes we needed to call findNode() to work out from which
node to initialise.

The solution is to map all complete tree nodes to some reduced tree node,
so that we can choose the mapping while we generate the reduced tree, and
make findNode() completely redundant.