Move nBlocks to the SmpsTree class.
The number of blocks depends on the tree, so it makes sense to store it in
the SmpsTree class, even though it's going to be used only by oops.

This allows to remove a call to orderNodes() that was necessary only to
recompute the number of blocks if a reduced tree was created.