Change get[Row|Col]Names() to accept an SmpsTree parameter.
This allows to generate correctly the names for a reduced tree, rather than
always generating names for the complete tree. This fixes the inconsistency
between allocation of names in get[Row|Col]Names() according to the complete
tree and their deallocation in ~SmpsReturn() according to the reduced tree,
which leaked.

This change is essential in case of decomposition, because there the root of
the reduced tree is no longer at period zero, and without it we would assign
the wrong set of names to rows and columns.