Be careful when changing the cutoff level.
The cutoff level needs to be adjusted in orderNodes() if the root node of
a tree is not at period zero. However, it has to be changed back otherwise
the generation of the complete problem fails.

This change restores the original cutoff level at the end of orderNodes();
for the reduced decomposed problems, the cutoff is adjusted in generateSmps()
as well.