Add a flag to denote when a warmstart point is ready to be used.
In case of decomposition, the allocated wsPoint is not be ready until all
the subproblems have been solved. Therefore, the check in setupProblem() for
wsPoint to be not NULL is wrong and may lead to crahses if decomposition is
used.

This adds the wsReady flag that is set to true when the warmstart point is
ready to be used as such.