Ensure in reduceTree() to not request more scenarios than are available.
We shouldn't ask for more scenarios than there are in the complete tree, as
otherwise reduceScenarios() would implicitly consider more stages than there
actually are.

The reason is that in this case rest is computed to be positive so we enter
the recursion on the child of the node. However, the node may not have any
children, so no further nodes are put in the reduced tree. The overall effect
is that reduceScenarios() will return a smaller reduced tree than desired.