Ignore the request to aggregate the last 1 stage.
This makes aggregateStages() execute only if we request to aggregate at
least two stages.

Note that before this change, "-a 1" was actually solving a reduced problem
with a tree corresponding to the complete tree, so it was really solving a
complete problem in two steps (loose tolerance first, and warmstarted tight
tolerance afterwards). Now, calling the solver with "-a 1" will run the cold
start version directly.