Set up is_col_diag of the precise size and reference its elements correctly.
In decomposition we have to work out the precise size of the [Rnk D0] block,
otherwise we end up generating it larger than necessary. Also the way of
referencing its elements, in the decomposition case, has to consider that
is_col_diag is indexed from zero but the first coreCol doesn't belong to the
first block.