Fix bug in the generation of the reduced tree.
In SmpsOops::reduceScenarios() we compute a step to move across the children
of a given node: this is used so that we do not take consecutive scenarios
for the reduced tree.

If the number of wanted scenarios is larger than the number of children, we
need to choose all children, so the step must be one. The bug was that in
this case the step was zero, which led to appending children nodes to the
previous parent, thus generating a reduced tree with the correct number of
scenarios but with the wrong shape.