Fix a segmentation fault when running the decomposition on watson.
The watson problems apply changes to the matrix. In decomposition, we may
end up with a change that requires access to the original root node (which
is not present in the decomposition subtree).

The fix is to avoid applying such changes, by checking that the column block
of the correction is not before the first period of the decomposition.