Avoid unnecessary work in SmpsDenseToVector().
SmpsDenseToVector() did not preserve the original ordering of the DenseVector
containing the warmstart point: instead, it would reorder it before copying
to the output vector and then reorder it back.

In the new version, SmpsDenseToVector() allocates a DenseVector that will
contain the reordered point from the beginning, so that the double reordering
of the original DenseVector can be avoided. This version does not consume
more memory than the previous one, since we now don't need to allocate a
work array in forwOrderColVector().