Store the root node of an tree from which a reduced is derived.
This adds a member variable to the SmpsTree class to store the root node of
the event tree from which the current one is derived. This is not NULL when
the tree is a reduced tree, an aggregated tree or a subtree.

The SmpsTree class doesn't have ownership of the orig node, so it must not
be freed in the destructor.